+++ score-http/src/main/java/org/oagi/score/repo/component/bdt_sc_pri_restri/BdtScPriRestriReadRepository.java
@@ -38,7 +38,7 @@ public class BdtScPriRestriReadRepository {
                 isDefault = "time".equalsIgnoreCase(xbtName);
             }
             availableBdtScPriRestri.setDefault(isDefault);
-            availableBdtScPriRestri.setXbtId(e.get(XBT.XBT_ID).toBigInteger());
+            availableBdtScPriRestri.setXbtId(e.get(XBT.XBT_ID));
             availableBdtScPriRestri.setXbtName(xbtName);
             return availableBdtScPriRestri;
         };
